# Make deadline of `NotifyIdle` configurable, allowing `flutter_smooth` to get 60FPS, even if GC needs to run for 14ms per 16.67ms