--- a/lib/ui/platform_dispatcher.dart
+++ b/lib/ui/platform_dispatcher.dart
@@ -523,6 +523,11 @@ class PlatformDispatcher {
   @FfiNative<Void Function(Bool)>('PlatformConfigurationNativeApi::SetNeedsReportTimings')
   external static void __nativeSetNeedsReportTimings(bool value);
 
+  void setAfterFrameNotifyIdle(Duration delta) => _setAfterFrameNotifyIdle(delta.inMicroseconds);
+
+  @FfiNative<Void Function(Int64)>('PlatformConfigurationNativeApi::SetAfterFrameNotifyIdleExtra')
+  external static void _setAfterFrameNotifyIdle(int delta);
+
   // Called from the engine, via hooks.dart
   void _reportTimings(List<int> timings) {
     assert(timings.length % FrameTiming._dataLength == 0);