--- a/runtime/runtime_controller.cc
+++ b/runtime/runtime_controller.cc
@@ -321,6 +321,11 @@ void RuntimeController::SetNeedsReportTimings(bool value) {
 }
 
 // |PlatformConfigurationClient|
+void RuntimeController::SetAfterFrameNotifyIdleExtra(fml::TimeDelta delta) {
+  client_.SetAfterFrameNotifyIdleExtra(delta);
+}
+
+// |PlatformConfigurationClient|
 std::shared_ptr<const fml::Mapping>
 RuntimeController::GetPersistentIsolateData() {
   return persistent_isolate_data_;