--- a/runtime/runtime_controller.h
+++ b/runtime/runtime_controller.h
@@ -628,6 +628,9 @@ class RuntimeController : public PlatformConfigurationClient {
   void SetNeedsReportTimings(bool value) override;
 
   // |PlatformConfigurationClient|
+  void SetAfterFrameNotifyIdleExtra(fml::TimeDelta delta) override;
+
+  // |PlatformConfigurationClient|
   std::unique_ptr<std::vector<std::string>> ComputePlatformResolvedLocale(
       const std::vector<std::string>& supported_locale_data) override;