--- a/runtime/runtime_delegate.h
+++ b/runtime/runtime_delegate.h
@@ -44,6 +44,8 @@ class RuntimeDelegate {
 
   virtual void SetNeedsReportTimings(bool value) = 0;
 
+  virtual void SetAfterFrameNotifyIdleExtra(fml::TimeDelta delta) = 0;
+
   virtual std::unique_ptr<std::vector<std::string>>
   ComputePlatformResolvedLocale(
       const std::vector<std::string>& supported_locale_data) = 0;