--- a/shell/common/animator.cc
+++ b/shell/common/animator.cc
@@ -250,7 +250,7 @@ void Animator::AwaitVSync() {
         }
       });
   if (has_rendered_) {
-    delegate_.OnAnimatorNotifyIdle(dart_frame_deadline_ + TODO);
+    delegate_.OnAnimatorNotifyIdle(dart_frame_deadline_);
   }
 }