@fzyzcjy 2022-11-05
> Same problems as the other NotifyIdle PR - these are details we don't want to expose to users.

Same confusion as in https://github.com/flutter/engine/pull/36797. Still think "dev tells system it can be idle for a period of time" is not something internal but a reasonable thing.

> NotifyIdle might not do anything, or it might do a lot more than the user expects.

Totally agree - I have checked Dart NotifyIdle source code and saw it uses something like history data to guess how long GC will happen now, but it can run for longer or shorter time.

As for this specific case of GC: If it does not do anything, there is surely no problem. If it do a lot more, it is still no problem. This is because, a much longer execution means that, even if we do not NotifyIdle, GC will happen somewhere in the future with at least this time duration (because garbage only accumulates).

It is also like what we do with `sleep`. Dart (and all lang) have a `sleep` function and can specify how long to sleep. But as you have pointed out earlier, the thread may not be woken up at the exact time because of OS schedule. Or, even consider this: Each and every line of Dart code (indeed Java/... as well), will have the possibility that, it takes much much longer time to finish! For example, needs 10ms to finish a single `i++`. This is just fact, not something that I make up - because we have the stop-the-world GC. When STW happens, the code just stuck. So, shall we ban `sleep` from users if the same logic holds, or even ban the whole Dart/Java/... languages ;) (Surely I am joking!)

> In all honesty, it would be nice to get rid of it entirely since it is hard to reason about and seems to come out wrong very frequently. If users can change this arbitrarily it will make it much harder to reason about what's going on in an application.

Sorry but do not get it.

May I know a bit more details why "wrong very frequently"? IMHO it just "notifies it is idle" and the only way to be wrong is the user forgets it can run for a much longer time - then my discussions about sleep and STW GC holds.

And I also wonder why "much harder to reason about what's going on in an application"? IMHO it is shown in the timeline, so it is quite clear what is going on. And this is indeed a Dart VM specific thing, so it is even not that related to Flutter framework and engine.